Move ReducingChatClient into library code
